bug fix
Because harvest can finish several days after it has started,
due to a stuck-utf8-filter problem or other problem, the
last-updated timestamp should be the harvest start date not
end date. For instance, if harvest starts on 1/1 and finishes
on 1/5, the last update date should be 1/1. If it was 1/5,
because incremental harvest starts from 1/5, updates between
1/1 and 1/5 won't be harvested.